Re-introduced caching in ConfigurationComponent
Analysis of the usage of ConfigurationService revealed that a complete
migration to Configurations#getConfiguration() is much effort and
may not always be desirable. The cache in the Configurations helper,
which was synchronized via EventService (see Configurations#onEvent()),
could become stale for short periods of time during a configuration
update. In critical cases, like the calculation of permits in
PermitService, this was unacceptable. On the other side, the
removal of the caching mechanism from ConfigurationComponent
imposed a performance penalty on callers of the readConfiguration()
method in ConfigurationService, because reading configurations from
storage is rather slow. For PermitService, which is time critical,
this was unacceptable, too.

Thus, this patch moves the caching of configuration values back to
ConfigurationComponent, and the Configurations helper has been
simplified considerably.

Both mechanisms to read configurations can now be used in parallel
and are served from the same cache. Hence, they should be equally
accurate and performant.

The Configurations helper should be used in code that is not an OSGI
service, or where a direct binding to the configuration service is
not desirable or not possible (e.g. because of cyclic dependencies).
Services may use both mechanisms, although Configurations should
be prefered to reduce the complexity of service binding during startup.

Change-Id: I2a08ae90b6818d14ec6ffff69e76e11e04e3612d
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>