Introduced event to notify about updated entities
Up to now in EntityServiceBase#persist() some additional "save actions"
were triggered, i.e. the update of the Lucene index and an asynchronous
validation with severity INFO. There was no possibility to register
more actions of that kind, such as update of a project cache or the
tag cloud.

Instead of adding an artifical "save action" mechanism to
EntityService, the eventing mechanism of Skalli has been reused.
A new event, EventEntityUpdate, has been added that is fired in
EntityServiceBase#persist() after persisting the entity in the
database. The event object provides the new entity value as parameter,
together with the id of the user that triggered the update.
Other services can listen to EventEntityUpdate events and react on
entity updates acoordingly.
This change requires servies derived from EntityService to bind
to EventService now in case they want to listen to update events
and register a corresponding listener. This has been implemented
for all services in o.e.s.core, especially for LuceneSearchComponent,
which updates the search index when a project is updated, and for
ValidationComponent, which triggers a re-validation.

The removal of the "save actions" from EntityServiceBase allowed
some simplifications, e.g. the activate/deactivate methods have been
removed etc., which caused some minor adaptions in some derived
classes.

Change-Id: I5d3df148c44ea399ce9a3296415972297c2bfd71
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>