TaggingService and TagCloud: Reimplemented
Up to now the tag clouds were calculated on the fly for the welcome
and the /tags pages, which worked quite well for small amounts of
projects. However, the algorithms in TaggingComponent and TagCloud were
non-linear in the number of projects, so the rendering time may become
unacceptable for large amounts of projects.

Since tags are read quite often but added/removed rather sporadically
a caching mechanism has been introduced in TaggingComponent that
listens to EventEntityUpdate events and re-indexes the tag cache
whenever a project is persisted. Read access for the tag clouds
is served now from the cache.

The TaggingService interface has been redesigned completely and all
the logic for calculating "most popular" tags and tag counts has been
moved from o.e.s.view.component.TagCloud to the tagging service.

The tag cache is designed to be fast and non-blocking on read access,
but is still thread-safe on updates.

Change-Id: Ic88dd30c9e436dae7b4dad0f358c364bc07b5d84
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>