Support for Gerrit 2.7 and beyond
Added 2.7 as recognized Gerrit version number. In case a version
number is provided by Gerrit but not in the known range, 2.4 is
assumed. This seems to be resonable, because 2.4 supports all
currently required features for Skalli.
This saves us from adjusting the version list for each new Gerrit
relase: Everything beyond 2.7 will just simply be treated as
compatible to 2.4.

Also added a new feature marker SUPPORTS_REST for versions >= 2.6.

Change-Id: I5c87cc0df3d49c2fd3e94f861052d0893b635fab
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>