[refactoring] Generic Feed Updater
SyndFeedUpdater was a quite generic utility to read arbitrary RSS
feeds from an URL, which should be the 90% case.

Therefore, SyndFeedUpdater was simplified and restructured, the API
documented and then moved to o.e.s.services.feed as URLFeedUpdater.
Functionality of the helper class Converter has been merged with
URLFeedUpdater, which was the sole user.

Some minor refactorings in o.e.s.model.ext.misc.

Change-Id: I97c819f1c46b7feac803185d80b27746e74bbd75
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>