[refactoring] XStream Persistence Implementation
XStreamPersistenceComponent:
Reordered the methods: moved methods implementing the interface to
the beginning and sorted them according to the declarations in the
interface. Added/improved some javadoc.

XStreamPersistence:
Reordered the method, moved the load/save methods called from
XStreamPersistenceComponent to the top. Reordered the helper methods.

XStreamPersistenceServiceTest:
renamed it to XStreamPersistenceComponentTest for consistency and
merged it with the mock class. Simplified and reformatted it

Cleaned up the formatting, did some minor renamings etc.

No functional changes!

Change-Id: I16c12c3f7332128bd353134912bb698bf5fda9df
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>