Entity Hierarchy: Introduced firstChild/nextSibling Pointers
Rendering the "All Projects" page with a large amount of projects,
or determining the subprojects of a given project in the REST API
suffered from the fact that a project knows only its parent project,
but not its children. While this is enough for storage, being able
to traverse the project hierarchy in both directions would be
desirable.

Thus, EntityBase got two new transient pointers:
firstChild : the first subproject of a project (or null)
nextSibling: the next sibling of a subproject (or null, if a
subproject is the last in the list of subprojects).

The calculation of the children is done in XStreamPersistenceComponent
during loading of the project model (now together with the parent
calculation).

Improvements for the "All Projects" page and REST API will come
in a subsequent patch.

Change-Id: I93b8542d6f78175f5df18f8b93a93df13cbc0018
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>