getSubProjects(): Fixed potential ConcurrentModificationException
Iterating through the result set of ProjectService#getAll() might
lead to a ConcurrentModificationException when underlying project
caches are changed during iteration (project modified/deleted):
better retrieve the UUIDs and copy them to a local variable, then
safely iterate on this copied list.

Note, the patch may change the order of the returned subprojects.
ProjectComponentPlace was depending on that order in one place,
which was wrong and has been fixed.

Moved assert helper method to AssertUtils and introduced a new one to
compare projects with a given UUID list in any order.

Change-Id: I4cf43f4e85c2feb1e0a3ef1a842658a20bd5066c
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>