[bug] XStreamPersistenceComponent#persist() destroys child pointers
loadEntity() does not resolve children of an entity, because that
would require a full scan through all projects. However, this led
to a problem when an entity loaded with loadEntity() was persisted
afterwards: the entity was put into the cache always with
firstChild==null, and therefore the children vanished from the UI
and from the REST API.

The fix ensures that the parent and children pointers of the entity
passed in to persist() are adjusted correctly now.

Note that the firstChild pointer will be forcefully overwritten
by persist() now to ensure that the parent/child relations always
stay in sync!

Added a test case for the persist-loadEntity-persist cycle.

Change-Id: I6f029a770a57aadee8c6c506fb4ed7f79f642af6
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>