[bug] NPE in GitGerritFilter when only a single Gerrit is configured
Dropdown for Gerrit selection is not rendered if only a single Gerrit
is configured, so that the corresponding form parameter "gerritId" is
undefined. Therefore gerritId==null, if form is submitted on "Check"
action, but when it is submitted from the "Check" page with "Proceed"
gerritId="" because the "Check" page contains a hidden input field
with value="".

Fixed the problem by checking if gerritId is null or blank, and in that
case the Gerrit server is not selected by id (which would return null),
but by GerritServerConfig#getPreferredServer(), which guarantees to
return a suitable Gerrit server.

Change-Id: I8609f16ffa93969ca101a3e55787c5ba4315ca2a
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>