[refactoring] Introduced UserService#getType()
Preparation for a subsequent patch: There is easy way to retrieve
properties from a service descriptor, so I added a getType() method
to UserService interface that allows clients to determine from
a given UserService instance which type of user store it supports.

Implemented getType() for the known user stores with the type
properties from the service descriptors.

Change-Id: I79ed6853ff9ea9efd3da84f5f38d08d54c23add0
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>