[refactoring] o.e.s.services.user.UserServices
Old implementation was not very efficient since the code for
determining the currently active user store was executed again
and again --- including a time consuming property lookup in other
bundles and service lookups for the configuration service and
the user service via a filter.

Refined UserServices analog to ExtensionServices/EntityServices
utilities: UserServices is registered as service, which binds to
UserService instances and caches them in a map by type.

Removed the now obsolete UserServicesTest.

Change-Id: Ieec0a4c060a61c1b82c3743cce270b4cbf130a92
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>