[refactoring]  Introduced RequestContext
Moved request parameters like action, path, query etc. from
ResourceBase to a new helper class RequestContext.

In a subsequent patch the media type will be added and we need to
pass the host, media type etc. to the new REST writers, so it makes
sense to extract them to a separate class.

However, ResourceBase still provides a facade for these parameters
for compatibility and convenience.

Change-Id: I158590f602632c868f2a418e2f82381e5c1bdadd
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>