[json]  Factory for RestWriters: RestService
The concrete RestWriter implementations should not be public.
Thus, we need kind of a factory to instantiate suitable RestWriters
for incoming REST requests, preferably as a service. Unfortunately,
JSONRestWriter and XMLRestWriter are not good services themselves
because they have an internal state. Consequently, a new service
interface has been introduced, o.e.s.services.rest.RestService,
that servers as factory for RestWriters.

RestService is implemented by o.e.s.core.rest.RestApplication, which
already is a service and seems to be the natural place for the
desired factory methods.

Change-Id: I552206a0cfcda6d2b5f24ea81b2e2d809a6fcf8f
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>