[json]  Preparation for RestConverter Migration
Introduced a new method to RestConverter,

    marshal(Object obj, RestWriter writer),

which is going to replace the inherited method from
com.thoughtworks.xstream.converters.Converter,

    marshal(Object source, HierarchicalStreamWriter writer,
            MarshallingContext context)

In this patch, both marshaling mechanisms co-exist, but everything
connected with the XStream mechanism has been deprecated and will
be removed once all converters have been migrated.

In order to migrate an existing converter, the abstract method

    protected void marshal(T obj)

defined in RestConverterBase must be implemented using the protected
"writer" attribute. Then, the corresponding Rest resource must be
migrated by switching the constructor.

Furthermore, ExtensionService#getRestConverter(host) has been
overloaded with a method without "host" parameter; the underlying
RestWriter knows the "host", so there is no need to pass it to the
RestConverter.

Change-Id: I0283e7293e420a4f9fa5e23e0f399194771a0ed2
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>