[bug] Child entity may point to outdated parent entity
Children of an "old" entity were not correctly assigned to a "new"
entity: The parent pointers of the children must be updated to
point to the "new" entity, but only the firstChild pointer of the
"new" entity was updated. This caused children pointing to an
outdated parent, if the parent was modified and persisted.

This bug was not detected in the tests, because one needs to check
for == of parent/child pointers, not equality of the parent/child
instances (which only ensures equality of the UUIDs).

Change-Id: If5a2abf6f4e50f91a53bacd568743f586e1daf46
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>