Better alert/message support
Welcome page was able to show exactly one alert and one info message.
However, an alert should be visible on all, or at least on the most
important pages; and there might be occasions, where one wants to
configure multiple alerts and info messages.

Changed the news configuration to support collections of alert
and informational messages. The welcome page now shows all alerts
and messages as usual in boxes below the tag cloud. Additionally,
searchheader.jsp that is included in all search and project detail
pages has been enhanced to display the alerts.

Change-Id: Ic0f4d9089206964cfacba1a1131c11a3bfd89515
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>