[bug] JiraFilter: Workaround for buggy WordUtils#abbreviate()
Some implemenatation of WordUtils#abbreviate() may cause a
StringIndexOutOfBoundsException in case the passed in string is
shorter than the specified "lower limit". However, in that case
it is not necessary to call abbreviate() at all, so the filter now
checks in advance whether the string already fits the desired
maximum length of 100 characters.

Change-Id: Ia2e9f32d2c590640826474d63a619d1215d12c20
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>