[refactoring] Property Cache for EntityBase
Determining the properties of an entity was quite expensive, because
the class had to be scanned for @PropertyName annotations, which
were applied to string constants and not to the actual getter methods.
Consequently, method like hasProperty() and getProperty() did only
check for the existence of a suitable getter method, ignoring the
@PropertyName annotations. Therefore, one could access properties
via REST API that were not actually qualifying as properties

This patch adds a lazily initialized property cache to EntityBase,
where properties together with their corresponding getter methods
are collected per class derived from EntityBase. Checks have been
added to hasProperty() and getProperty() ensuring that a requested
property has a corrresponding string constant annotated with
@PropertyName and a suitable getter method. Nice side effect:
because the getter methods are collected once and cached, performance
of these methods should improve, too.

Change-Id: I98746a60515d61470d284fbb049c1caff733d7aa
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>