Property Accessor Methods
The classical @PropertyName mechanism was only capable of providing
simple and collection-like properties, but for example not elements
of map-like properties, or inner properties of a property.

This patch adds a new annotation @Property that can be applied to
arbitrary methods following the pattern

  @Property Object getProperty(String...params)

Moreover, EntityBase got a new method getProperty(Expression...),
that allows do drill down into the properties of an entity with the
help of such "property acccessor" methods.

Change-Id: Iabb74616a5027da57b0ad7b7771f2feb38dc0193
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>