[rest] /api/projects/: support for enhanced property queries
Adapted the match algorithm to support queries of the form

  /api/projects?properts=ext.prop(name).item(key)&pattern=...

implemented by property accessor methods.

In the example, the extension with the short name "ext" must have
a method with the signature "Object getProp(String)" that is annotated
with @Property, and the returned object must implement a method with
the signature "Object getItem(String)" that is also annotated with
@Property. The result of getItem() is compared with the given pattern
after conversion to a string.

Queries of the form "property=ext.propertyName" are still supported
and can be freely mixed with the more complex queries, e.g.

   /api/projects?properts=ext.prop(name).value&pattern=...

would require the result of getProp() to have a classical property
"value" that is defined by a string constant annotated with
@PropertyName.

Change-Id: I825d35cf4b4b54e16e76decafc8440ab49996f5a
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>