[rest] Support "extensions" attribute for /api/projects/{id}
Adding an "extensions" attribute to an /api/projects/{id} request
with a comma-separated list of short names just returns the details
of the extensions matching these short names. The default remains
returning all extensions, in contrast to /api/projects, which returns
no extensions at all unless an "extensions" attribute is specified.
The result for requests with "extensions" attribute, however, will be
the same for /api/projects and /api/projects/{id}.

Change-Id: Ifecd9ab7cdc284557d6711b5f268c83ad2b8a434
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>