[rest] RestService#getRestWriter(): pass RequestContext as parameter
All parameters (except Writer) of this method are actually parameters
of the current request context, so instead of passing each of them
individually and changing the interface each time we need a new one,
we could just use the whole request context as parameter.

This enhanced flexibility will be explored in a subsequent patch
to provide a new query parameter, "prettify", which will switch on
a prettifying converter for JSON.

Since this is an incompatible change, some tests had to be adapted,
too. In the course of fixing the tests I moved a nice helper method
from ErrorRepresentationTest to the base class RestWriterTestBase
and reused it for all the tests.

Change-Id: I5c27d6ebfea2b53c3bb681d6dababf23086c251e
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>