[rest v2] Avoid ugly duplicate keys in JSON output
Extensions are rendered as JSON objects, because we have to render
some common attributes like "apiVersion" and "inherited". However,
this can lead to ugly duplications of element keys when the payload
of an extension is merely an array.

The "tags" extension, for example, up to now was rendered as:

{ "tags": {
  "apiVersion":"1.0",
  "tags": ["tag1","tag2]
  }
}

In situations like this we now use the key "items" by convention,
which corresponds to common JSON style guides, i.e.

{ "tags": {
  "apiVersion":"1.0",
  "items": ["tag1","tag2]
  }
}

The new convention has been applied to all converters which used the
same key for the outmost element (usually an object) and for a direct
child array.

Change-Id: I04a2ecc03f1815806ae32623675510975f78b24c
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>