[rest] Added proper handling of unsupported media type
REST API failed with 500 Internal Server Error in case of unsupported
media type in the request. Reason was that the check for the media
type was done too late, i.e. in ResourceRepresentation#write(). At
this point in processing there was no other possibility than to
throw an IOException, which was caught by the RESTlet container.

This patch adds a check for the media type at the beginning of
all GET handlers of all known REST resource implementations. If
the media type is unknown a proper 415 Unsuppported Media Type
response is now produced.

ResourceBase#createUnsupportedMediaTypeRepresentation() has been
removed, because if the media type is unsupported this method
inevitably will throw an IOException, too.

Change-Id: I97eab832c25b6ea9d4dae6bc110e1c9d0c23583c
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>