Revert source/target compliance back to JDK 7
It turned out that with recent JDK 7 implementations the Tycho
build fails with a java.lang.IllegalArgumentException in the p2
mirror selector because of a violation of the Comparable interface
contract. This bug has been fixed in newer Equinox versions.
Means: We need to migrate to a newer Tycho version. However, with
newer Tycho versions all surefire tests break because of an
incompatible change of the dependency resolution mechanism in Tycho.
Furtermore, newer Equinox versions are not compatible with the
rather ancient Jetty we use for building the standalone product.
Migrating to a newer Jetty is unfortunately not an easy one since
the Jetty project no longer ships the OSGI-enabled variant Jetty-RT
that we need for building the standalone product. Gotcha.

Before building with JDK 7 can be enabled, the whole build and
runtime chain has to be rebuilt from scratch.

Change-Id: I2bb934b8736462b4a9cb94fd5e40c55a063c3798
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>