[releng] Reorganize Orbit and Eclipse Features
Moved all dependencies that correspond to the Orbit section in
the target platform to the Orbit feature. Moved all remaining
Eclipse dependencies to the Eclipse feature.

Sorted both feature.xml alphabetically and replaced the 0.0.0
version placeholders with concrete versions to detect mismatches
and missing dependencies more easily in the future.

Change-Id: I91271fddc44beeac910c4041d93ef5569610efee
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>