Adapt rules for compilation of JSP files
By default, Jasper runs in development mode, i.e. checks for
each request if JSP files have changed and must be recompiled.
That is not appropriate for a productive environment.

This patch adds a configuration section for the JspServlet to the
web.xml of o.e.s.view. By default "reload-interval=-1" is set now,
i.e. JSP files are compiled on first hit, but then never checked
for changes again. Other useful parameter combinations are given
in the web.xml, too, but are commented.

Change-Id: I2d0e63b82389acb88fb457a68626b363995de5a9
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>