Ensure that JPA 2.0 is available and actually used
o.e.s.jpa and o.e.s.api have a dependency to JPA 2.0, i.e.
TypedQuery, which is not available in earlier JPA releases.
If the javax.persistence package is available as JPA 1.0 and
JPA 2.0 Tycho tends to take JPA 1.0, which breaks the build.

Added jpa:2.0 modifiers to the javax.persistence manifest entries
of the mentioned bundles.

Change-Id: If892a46b1bcd30e9c17a0d2733319af02224ad06
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>