Build standalone product only if build-product profile active
Build of the the standalone product and the necessary
orbit/jetty/eclipse features is now performed only, if Tychco is
started with the profile parameter -Pbuild-product.
By default, only a p2 repository with the Skalli bundles,
source bundles and testutils bundle is produced.

This speeds up the build considerably and simplifies consumption
of Skalli for use in other web containers.

Change-Id: Ia47a8703a4b6e8d4940fa81a7e13f7fa4780da2c
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>