New profile "code-coverage"
This patch adds a profile "code-coverage" to the reactor POM, 
which configures the Jacoco VM agent. The Jacoco plugin sets
the property "tycho.testArgLine" that is passed as additional
argument to the tycho-surefire plugin during test execution.
The output of the Jacoco agent is written to /target/jacoco.exec.
All test bundles report to the same file.

For symmetry the property for custom surefire arguments has been
renamed to "testArgLine".

Change-Id: Ifb9329f38f89df8d48bfffb8809c16ce515d9086
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>