Missing tycho.testArgLine if only "static-checks" profile is active
A profile with activeByDefault=true will not become active if a
distinct profile has been requtested in the Maven call. This broke
the build if any profile was activated without activating the
"code-coverage" profile at the same time.

This patch changes the activation mechanism of the profiles
"static-checks" and "code-coverage". Instead of activating by
profile id a corresponding property must be passed to the Maven
call, i.e.

  -Dstatic-checks   activates the findbugs and PMD checks
  -Dcode-coverage   activates the jacococ code coverage analysis 

All other profiles were not touched.

Change-Id: I6557ac8922f3f37ce60bcde433b07929611296ca
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>