Enable RestAPISmokeTest
This test was switched off because with the permission handling
in place it did no longer work. There is no way to circumvent
the login procedure, which is not compatible with the built-in
server provided with org.restlet.

This patch employs an embedded Jetty server providing a servlet
container for running RestletServlet. This way the whole web chain
is tested and not only the org.restlet part.

Change-Id: I04013997bf30e65f59b8d7f70ae285013bcf2626
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>