Employ XMLDiffUtil to test output of REST API
This patch adds files containing the expected results of the tests
as bundle resources and employs the XMLDiffUtil to compare the
actual output of the REST API with these files (after replacing
a placeholder for the port, which may vary from test to test).

Since all the tests basically do the same now the common parts of
the tests have been extracted to a helper method.

Change-Id: I8eb83b610b4043019fa40233495c1cddc826a3ce
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>