Support for "orderBy" query parameter for /api/projects
The supported sort orders are defined in the enum SortOrder.
Currently supported values for the query parameter are "uuid",
"projectId" and "name" (all case-insenitive). Any other value,
or if the parameter is no given at all, no sorting should be
applied to the result of a Lucene query (or any other set of
projects).

The parameter is ignored when also a "query" parameter is
specified. In that case the sort order is determined by
Lucene. Otherwise the default sort order is undefined as
was before this patch.

Change-Id: I20f7b21345d6f0143ccbaea997114cf786561ee6
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>