Unit tests for comparing v1 and v2 XML REST output
This patch adds a test that connects to a running Skalli instance
specified with a system property "skalli.itest.host", for example
http://localhost:8080, iterates over all projects in chunks of
10 projects and compares the XML response produced by the old (v1)
and the new (v2) REST writers. Since the output is expected to
be slightly different, a dedicated XMLUnit DfferenceListener must
be applied.

Note, for running this test in Tycho, add something like

  -Dskalli.itest.host=http://localhost:8080

to the <testArgLine> property in the reactor POM because surefire
tests are executed in a separate process.

Change-Id: I0296f468a4b055fb1eedd3d6597e9c7b98d02d70
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>