Unit tests for comparing v1 and v2 XML REST detail output
Enhancements for CompareAPIVersionTest to compare also the detail
information of projects (/api/projects/<uuid>). The test still
downloads and compares chunks of 10 projects from the all-projects
list (/api/projects), but then iterates through the result list
and downloads/compares the output of the v1 and v2 project detail
writers (CommonProjectConverter).

Note, this test may not work on a Skalli instance with additional
extensions.

Change-Id: I466a377b6aaef53758b82f04b5c0cf2fc969accc
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>