Introduce InheritableExtension
This is a wrapper class for ExtensionEntityBase plus an "inherited"
flag. This is necessary since in the current model inheritance is
tracked by the project. The "inherited" flag is an instance of
Boolean so that one can express the three states "inherited",
"not inherited" and "undetermined". The latter will be used to
switch inheritance on/off in the REST API.

Change-Id: I1aefdbd308041e9f2318a1fa97d7b7d5447ad540
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>