Introduce ExtensionService#newExtension()
For type-safe implementation of the write REST API we need to
be able to create new extension instances with initial values.
We could, or course, use Class#newInstance(), but that might
not always be possible or correct (note the exceptions that
newInstance() might throw!).

The natural place for creating new instances is in the
corresponding extension services, so this patch introduced a
new method ExtensionService#newExtension() and a default
implementation in ExtensionServiceBase to retain compatibility.
Additionally, for all known extension services this patch
provides type-safe implementations of newExtension().

Note that custom extension services are explicitly encouraged
to overwrite the default implementation of newExtension(), too!

Change-Id: I5c2b7b943d8a6b38d328af6cebce719739d09ea9
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>