Handle inherited extension with extra attributes gracefully
In case the "inherited" flag was set on an inheritable extension
unmarshalling always failed on the closing reader#end() call
with a parsing exception because the following common attributes
and/or extension attributes were never consumed. While that
behavior is, strictly speaking, correct throwing a parsing
exceptions seems unnecessarily harsh in that case.

This patch adapts InheritableExtensionConverter to always parse
all available attributes and return them to the caller of the
unmarshall method. The caller then may decide how to handle
the result, e.g. produce a less technical error message or just
ignore it.

Change-Id: I7935c00129457b0f185edb92f4ad2581a106c8b5
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>