[refactoring] Simplify CommonProjectConverter constructors
CommonProjectConverter had a lot of flags to tweak the rendering
of extensions, namespaces and common attributes like lastModified.

This patch removes all the flags and moves the rendering of
the namespace and common attributes to the derived classes,
e.g. to ProjectConverter, and to the callers of the constructors,
e.g. ProjectsCionverters.

There are now 2 constructors remaining:

  CommonProjectConverter()

The project will be rendered without extensions.

  CommonProjectConverter(String[] extensions)

Renders the project with the given extensions. If the argument
is null, no extensions will be rendered. If the argument array
contains an "*" entry, all extensions will be rendered.

Change-Id: Ia7d8116f557079ccfc2800c06fbb54ef0063bd66
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>