Re-introduce Project#getParentProject() returning Project
There were several places in the code where the result of
EntityBase#getParentEntity() was cast to Project, because
a type-safe "getParentProject()" method was not available
for Project.The old method of that name returned only an
UUID.

This patch re-introduces Project#getParentProject(),
but now with return type Project. The necessary cast
from EntityBase to Project is performed with a checked
cast, i.e. with Class#cast().

Change-Id: If7fbac9d7b6796c74d295a3d65b81169123e5f09
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>