Enrich the "parent" link in project REST output
This patch adds uuid, id and name attributes to the
parent link of a project's REST output similiar to
the subproject links. These additional attributes,
for example, allow rendering a project hierarchy in
a UI presenting logical/display names without the
need for an additional REST request for the parent
project.

Change-Id: I3944fb51c1939e68f3e65ec9c552ac83021fb66a
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>