[rest] Support for creating projects
POST /api/projects is introduced that allows creating and
initializing new projects. The symbolic name of the project
must be specified with a Slug header.

If a request body is available the project is initialized
from the request. Otherwise an initial project with just the
uuid and symbolic name is created.

Change-Id: I7622986b50771c1cdf23ce82011c9c4dfaaa89c2
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>