ErrorRepresentation fails with Unsupported Media Type
There are some REST requests supporting PUT operations with
media type text/xml (e.g. the backup REST API). In that case
ErrorRepresentation failed with an IOException stating that
the media type text/xml was not supported for PUT operations.
However, that is highly irrelevant, because writing an error
response is a GET operations.

This patch removes the redundant check for the media type of
the original request.

Change-Id: I52c34b94e059e95cc2c1a4209f8b6c687034b6e0
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>