Replace usage of org.json with com.google.gson
It makes not much sense to use different JSON frameworks in the
REST API and the Gerrit client. Since com.google.gson provides
better streaming capabilities, we settle on com.google.gson.

This patch removes org.json from the target platform and manifests
and moves the re-implemented JSONUtil to the org.eclipse.commons
package. The name is now JSONUtils, and it has been enriched
with some new utility methods.

Change-Id: Ic11d91d0e628d9cecb146828ed113e337df2819a
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>