Remove overly pessimistic locking in ConfigurationComponent
readConfiguration() and writeConfiguration() both were declared
as synchronized although both read/write data from/to a cache,
which is a ConcurrentHashMap. That was two pessimistic and
caused a deadlock at least once in LDAPUserComponent.

This patch removes the synchronization from the mentioned methods
assumingh that the underlying storage service is thread safe, which
every service must be. It's not true for the existing implementations
in org.eclipse.skalli.core. That should be fixed in a subsequent
patch since it needs some refactoring first.

Change-Id: Ia3f430829f9b1a40d49bda7403c0f01c6d3587a1
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>